// client/src/statemanagement/SoknadStatus.ts
export enum SoknadStatus {
  VENTER_GODKJENNING = 'VENTER_GODKJENNING',
  GODKJENT = 'GODKJENT',
  SLETTET = 'SLETTET',
  GODKJENT_MED_FULLMAKT = 'GODKJENT_MED_FULLMAKT',
  INNSENDT_FULLMAKT_IKKE_PÅKREVD = 'INNSENDT_FULLMAKT_IKKE_PÅKREVD',
  UTLØPT = 'UTLØPT',
  ENDELIG_JOURNALFØRT = 'ENDELIG_JOURNALFØRT',
  VEDTAKSRESULTAT_INNVILGET = 'VEDTAKSRESULTAT_INNVILGET',
  VEDTAKSRESULTAT_MUNTLIG_INNVILGET = 'VEDTAKSRESULTAT_MUNTLIG_INNVILGET',
  VEDTAKSRESULTAT_DELVIS_INNVILGET = 'VEDTAKSRESULTAT_DELVIS_INNVILGET',
  VEDTAKSRESULTAT_AVSLÅTT = 'VEDTAKSRESULTAT_AVSLÅTT',
  VEDTAKSRESULTAT_HENLAGTBORTFALT = 'VEDTAKSRESULTAT_HENLAGTBORTFALT',
  VEDTAKSRESULTAT_ANNET = 'VEDTAKSRESULTAT_ANNET',
  UTSENDING_STARTET = 'UTSENDING_STARTET',
  BESTILLING_AVVIST = 'BESTILLING_AVVIST',
  BESTILLING_FERDIGSTILT = 'BESTILLING_FERDIGSTILT'
}